#14 Added implementation to appropriate Activity class
