+++ SmartCarMQTTApp/app/src/main/AndroidManifest.xml
@@ -15,9 +15,6 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.SmartCarMQTTApp">
         <activity
-            android:name=".ControlActivity"
-            android:exported="false" />
-        <activity
             android:name=".PracticeTheoryActivity"
             android:exported="false" />
         <activity
