+++ SmartCarMQTTApp/app/src/main/java/com/example/smartcarmqttapp/MainActivity.java
@@ -37,7 +37,7 @@ public class MainActivity extends AppCompatActivity {
         new Handler().postDelayed(new Runnable() {
             @Override
             public void run() {
-                Intent in = new Intent(MainActivity.this, ControlActivity.class);
+                Intent in = new Intent(MainActivity.this, HomeActivity.class);
                 startActivity(in);
                 finish();
             }
